// src/shared/enums/component-custom-id.ts
export enum ComponentCustomIdEnum {
  TIME_MODAL = 'TIME_MODAL',
  SET_UTC_MODAL = 'SET_UTC_MODAL',
  MODAL_INPUT_NICKNAME = 'MODAL_INPUT_NICKNAME',
  MODAL_INPUT_NICKNAME_FOR_EDIT = 'MODAL_INPUT_NICKNAME_FOR_EDIT',
  INPUT_NICKNAME = 'INPUT_NICKNAME',
  INPUT_NICKNAME_FOR_EDIT = 'INPUT_NICKNAME_FOR_EDIT',
  ELEMENT = 'ELEMENT',
  OPEN_PANEL_CREATE_PARTY = 'OPEN_PANEL_CREATE_PARTY',
  OPEN_PANEL_CHARACTER_MANAGER = 'OPEN_PANEL_CHARACTER_MANAGER',
  OPEN_PANEL_EDIT_CHARACTER = 'OPEN_PANEL_EDIT_CHARACTER',
  TOGGLE_STATUS = 'TOGGLE_STATUS',
  SELECT_SERVER = 'SELECT_SERVER',
  SELECT_CLASS = 'SELECT_CLASS',
  SELECT_GENERAL_CLASS = 'SELECT_GENERAL_CLASS',
  SELECT_ELEMENT = 'SELECT_ELEMENT',
  SELECT_NEST = 'SELECT_NEST',
  SELECT_CHARACTER = 'SELECT_CHARACTER',
  SELECT_CHARACTER_FOR_EDIT = 'SELECT_CHARACTER_FOR_EDIT',
  SELECT_CHARACTER_FOR_CREATE_PARTY = 'SELECT_CHARACTER_FOR_CREATE_PARTY',
  CLASS_PRIORITY_LOOT_TOGGLE = 'CLASS_PRIORITY_LOOT_TOGGLE',
  TIME_START = 'TIME_START',
  TIME_END = 'TIME_END',
  SET_UTC_MODAL_INPUT = 'SET_UTC_MODAL_INPUT',
  OPEN_MODAL_SET_TIME = 'OPEN_MODAL_SET_TIME',
  OPEN_MODAL_INPUT_NICKNAME = 'OPEN_MODAL_INPUT_NICKNAME',
  OPEN_MODAL_INPUT_NICKNAME_FOR_EDIT = 'OPEN_MODAL_INPUT_NICKNAME_FOR_EDIT',
  OPEN_MODAL_INPUT_UTC_FOR_CREATE_PARTY = 'OPEN_MODAL_INPUT_UTC_FOR_CREATE_PARTY',
  SUBMIT_CREATE_PARTY = 'SUBMIT_CREATE_PARTY',
  SUBMIT_CHARACTER_ADD = 'SUBMIT_CHARACTER_ADD',
  SUBMIT_CHARACTER_DELETE = 'SUBMIT_CHARACTER_DELETE',
  SUBMIT_CHARACTER_EDIT = 'SUBMIT_CHARACTER_EDIT',
  CHECK_MORE_INFO_ABOUT_UTC = 'CHECK_MORE_INFO_ABOUT_UTC',
  CREATE_PARTY_MOVE_TO_STAGE_2 = 'CREATE_PARTY_MOVE_TO_STAGE_2',
}